Consolidate App and Main. Expose websocket errors.
